+++ README.md
@@ -31,25 +31,30 @@ mvn clean package
 **Help**
 
 ```
-java -jar noark-extraction-validator-0.3.0.jar
+java -jar noark-extraction-validator-0.4.0.jar
 ```
 OR
 ```
-java -jar noark-extraction-validator-0.3.0.jar --help
+java -jar noark-extraction-validator-0.4.0.jar --help
 ```
 
 **Validating different extraction package versions**
 
-Currently supported versions are *Noark 5.3* and *Noark 5.4*.
+Currently supported versions are *Noark 5.3*, *Noark 5.4* and *Noark 5.5*.
 
 Noark 5.3:
 ```
-java -jar noark-extraction-validator-0.3.0.jar noark53 ...
+java -jar noark-extraction-validator-0.4.0.jar noark53 ...
 ```
 
 Noark 5.4:
 ```
-java -jar noark-extraction-validator-0.3.0.jar noark54 ...
+java -jar noark-extraction-validator-0.4.0.jar noark54 ...
+```
+
+Noark 5.5:
+```
+java -jar noark-extraction-validator-0.4.0.jar noark55 ...
 ```
 
 To validate against the desired version, substitute the `<noark-version>` parameter in the following examples with one of the supported versions.
@@ -58,7 +63,7 @@ To validate against the desired version, substitute the `<noark-version>` parame
 
 If you would like to enable logging in the console or in a specific file, you will need to provide a reference to a log4j2 configuration:
 ```
-java -Dlog4j.configurationFile='/path/to/log4j2.xml' -jar noark-extraction-validator-0.3.0.jar <noark-version> ...
+java -Dlog4j.configurationFile='/path/to/log4j2.xml' -jar noark-extraction-validator-0.4.0.jar <noark-version> ...
 ```
 A default configuration exists in a directory called `config/`.
 
@@ -66,26 +71,26 @@ A default configuration exists in a directory called `config/`.
 
 The shortest command to validate a package is:
 ```
-java -jar noark-extraction-validator-0.3.0.jar <noark-version> -extraction /path/to/uttrekk/directory
+java -jar noark-extraction-validator-0.4.0.jar <noark-version> -extraction /path/to/uttrekk/directory
 ```
 More complex scenarios would include modifying the output report type and/or directory:
 ```
-java -jar noark-extraction-validator-0.3.0.jar <noark-version> -extraction /path/to/uttrekk/directory -output-dir /path/to/report/output/directory -output-type excel_xls
+java -jar noark-extraction-validator-0.4.0.jar <noark-version> -extraction /path/to/uttrekk/directory -output-dir /path/to/report/output/directory -output-type excel_xls
 ```
 You can request a report to be generated in multiple formats:
 ```
-java -jar noark-extraction-validator-0.3.0.jar <noark-version> -extraction /path/to/uttrekk/directory -output-dir /path/to/report/output/directory -output-type excel_xls -output-type xml
+java -jar noark-extraction-validator-0.4.0.jar <noark-version> -extraction /path/to/uttrekk/directory -output-dir /path/to/report/output/directory -output-type excel_xls -output-type xml
 ```
 You can specify custom Noark schemas to be used in the validation process (in addition to the Noark ones and the extraction package ones). In such cases the reports would also include information about the compliance of the extraction package to these schemas:
 ```
-java -jar noark-extraction-validator-0.3.0.jar <noark-version> -extraction /path/to/uttrekk/directory -custom-schema-location /path/to/custom/schemas/directory
+java -jar noark-extraction-validator-0.4.0.jar <noark-version> -extraction /path/to/uttrekk/directory -custom-schema-location /path/to/custom/schemas/directory
 ```
 The specified `-custom-schema-location` directory may contain any custom Noark 5 schemas and a UTF-8-encoded description.txt file. The contents of this file will be copied to the Execution Information sections of the generated reports for completeness.
 
 
 You can also change the persistence settings:
 ```
-java -jar noark-extraction-validator-0.3.0.jar <noark-version> -extraction /path/to/uttrekk/directory -output-dir /path/to/report/output/directory -db-name myDBName -storage hsqldb_server -server-location http://my.hsqldb.server.com
+java -jar noark-extraction-validator-0.4.0.jar <noark-version> -extraction /path/to/uttrekk/directory -output-dir /path/to/report/output/directory -db-name myDBName -storage hsqldb_server -server-location http://my.hsqldb.server.com
 ```
 *Do not modify* the persistence settings unless:
 * You would like to run several instances of the validator at the same time
@@ -99,7 +104,7 @@ java -jar noark-extraction-validator-0.3.0.jar <noark-version> -extraction /path
 
 Making the validation process less strict is also possible. The default behavior of the validator is to stop execution if an XML file does not comply with the corresponding Noark XSD schema. To continue execution instead:
 ```
-java -jar noark-extraction-validator-0.3.0.jar <noark-version> -extraction /path/to/uttrekk/directory -ignore-non-compliant-xml
+java -jar noark-extraction-validator-0.4.0.jar <noark-version> -extraction /path/to/uttrekk/directory -ignore-non-compliant-xml
 ```
 This, however, is a bad practice that we would *advise strongly against* - the reason being that the result of the validation in such cases may be unreliable because crucial metadata elements are missing.
 
@@ -108,7 +113,7 @@ This, however, is a bad practice that we would *advise strongly against* - the r
 Similar to the above but note that Windows uses a different path separator:
 
 ```
-"c:\Program Files\Java\jdk1.8.0_111\bin\java" -jar noark-extraction-validator-0.3.0.jar <noark-version> -extraction c:\path\to\uttrekk -output-dir c:\path\to\output\reports\dir
+"c:\Program Files\Java\jdk1.8.0_111\bin\java" -jar noark-extraction-validator-0.4.0.jar <noark-version> -extraction c:\path\to\uttrekk -output-dir c:\path\to\output\reports\dir
 ```
 
 Also note that the above command specifies the path to Java explicitly to make sure that we are using a JDK, not a JRE (if one is installed). There are other ways to achieve the same thing.
@@ -127,6 +132,9 @@ Noark53Command command = new Noark53Command();
 
 # To create a Noark 5.4 command, use alternatively:
 # Noark54Command command = new Noark54Command();
+
+# Similar for Noark 5.5:
+# Noark55Command command = new Noark55Command();
 
 command.setExtractionDirectory(new File("/path/to/extraction"));
 command.setReportConfiguration(reportConfiguration);
