+++ noark-extraction-validator/src/main/java/com/documaster/validator/validation/ValidationFactory.java
@@ -20,9 +20,11 @@ package com.documaster.validator.validation;
 import com.documaster.validator.config.commands.Command;
 import com.documaster.validator.config.commands.Noark53Command;
 import com.documaster.validator.config.commands.Noark54Command;
+import com.documaster.validator.config.commands.Noark55Command;
 import com.documaster.validator.exceptions.ValidationException;
 import com.documaster.validator.validation.noark5.noark53.Noark53Validator;
 import com.documaster.validator.validation.noark5.noark54.Noark54Validator;
+import com.documaster.validator.validation.noark5.noark55.Noark55Validator;
 
 public class ValidationFactory {
 
@@ -33,6 +35,8 @@ public class ValidationFactory {
 				return new Noark53Validator((Noark53Command) command);
 			case NOARK54:
 				return new Noark54Validator((Noark54Command) command);
+			case NOARK55:
+				return new Noark55Validator((Noark55Command) command);
 			default:
 				throw new ValidationException("Validator not implemented: " + validatorType.getName());
 		}
