+++ noark-extraction-validator/src/main/resources/noark5/noark55/noark55.properties
@@ -1,0 +1,7 @@
+# Noark 5.5 package checksums
+schemaChecksums.arkivstruktur.xsd = 85986f7c8fac408cca568a0436b26f5b2837d420877a529962437d21c71fac82
+schemaChecksums.metadatakatalog.xsd = df9c4bb29a4fc49d452586337c01666f286c7070fe170792b99faf3ef652cf8c
+schemaChecksums.addml.xsd = 4dc20e5dbf21e94cf211464e9a60b99acd24c45b46ecbf872a32fe2384b904e6
+schemaChecksums.endringslogg.xsd = 9c0aa09d77ce76077f6f65f0a7500f5cf3dadd82410a1ab38cbfd32a76e93d20
+schemaChecksums.loependeJournal.xsd = e1c2737159f40e67627329de851d49d841617dddc360ccf3aeea1ffc1ef72e6c
+schemaChecksums.offentligJournal.xsd = 0e8d3b49eeb9eab496bc2ea2f8af9160e6e2526485fe27e377a508b3946bff6e
